Rename /hocs to /contexts
