+++ plugin-hrm-form/src/components/search/SearchForm/index.jsx
@@ -6,7 +6,7 @@ import FieldText from '../../FieldText';
 import FieldSelect from '../../FieldSelect';
 import FieldDate from '../../FieldDate';
 import { Container, StyledNextStepButton, Row, BottomButtonBar } from '../../../Styles/HrmStyles';
-import { withConfiguration } from '../../../hocs/ConfigurationContext';
+import { withConfiguration } from '../../../contexts/ConfigurationContext';
 import { configurationType, searchFormType } from '../../../types';
 
 const getField = value => ({
